Split code into a lib crate for a better modularity ✂️
